chore: Better structure for commands
