+++ src/bots/cmd_handler.py
@@ -11,7 +11,6 @@ from data import data
 from discord import Embed
 
 
-
 commands = {}
 
 def command(discord=False, tfm=False, whisper_command=False, aliases=None, allowed_roles=None):
@@ -29,6 +28,15 @@ def command(discord=False, tfm=False, whisper_command=False, aliases=None, allow
 
         return wrapper
     return decorator
+
+
+from .commands import ccmds
+
+command(discord = True)(ccmds.cmds)
+command(discord = True)(ccmds.cmd)
+command(discord = True, allowed_roles = [ data["roles"]["admin"], data["roles"]["cmder"] ])(ccmds.ccmd)
+command(discord = True, allowed_roles = [ data["roles"]["admin"], data["roles"]["cmder"] ])(ccmds.ecmd)
+command(discord = True, allowed_roles = [ data["roles"]["admin"], data["roles"]["cmder"] ])(ccmds.dcmd)
 
 
 @command(discord=True)
@@ -231,88 +239,3 @@ async def profile(args, msg, client):
             embed["image"] = { "url": "http://santanl.000webhostapp.com/badges.php?row_max=18&badges={}".format(",".join([str(x) for x in tfm_profile.badges.keys()])) }
 
         await msg.reply(embed = Embed.from_dict(embed))
-
-@command(discord=True)
-async def cmds(args, msg, client):
-    await msg.reply(embed = Embed.from_dict({
-        "title": "Available custom commands",
-        "description": "• " + "\n• ".join(client.ccmds.keys()),
-        "color": 0x2987ba
-    }))
-
-@command(discord=True)
-async def cmd(args, msg, client):
-    cmd_name = utils.get(args, 0, "")
-    cmd = client.ccmds.get(cmd_name)
-    if not cmd:
-        return await msg.reply(":x: | Cannot find that command")
-    await msg.reply(embed = Embed.from_dict({
-        "title": "Custom command info",
-        "fields": [
-            { "name": "Author", "value": "<@!{}>".format(cmd["author"]) },
-            { "name": "Source", "value": "[{source}]({source})".format(source = cmd["source"])}
-        ],
-        "color": 0x2987ba
-    }))
-
-@command(
-    discord=True,
-    allowed_roles = [ data["roles"]["admin"], data["roles"]["cmder"] ]
-)
-async def ccmd(args, msg, client):
-
-    #admin_role = client.main_guild.get_role(data["roles"]["admin"])
-    #cmder_role = client.main_guild.get_role(data["roles"]["cmder"])
-
-    #if (not admin_role in msg.author.roles) or (not cmder_role in msg.author.roles):
-    #    return await msg.reply(embed = Embed.from_dict({
-    #        "title": ":x: Error",
-    #        "description": "You need the <@&{}> role to manage commands".format(data["roles"]["cmder"]),
-    #        "color": 0xcc0000
-    #    }))
-
-    if len(args) < 3:
-        return await msg.reply(":x: Failed to create the command. Please supply all the arguments\nFormat: `> ccmd <name> <compiler> <source>`")
-    name, compiler, source = args[0], args[1], args[2]
-    if len(name) > 10:
-        return await msg.reply(":x: Command name should be less than or equal to 10 characters")
-    if client.ccmds.get(name):
-        return await msg.reply("Command **{}** already exists! Please use `> ecmd` to overwrite it!".format(name))
-
-    client.ccmds[name] = { "runner": compiler, "source": source, "author": str(msg.author.id) } # stringified id for backward compatibility
-    await client.update_ccmds()
-    await msg.reply(":white_check_mark: | Created the command")
-
-
-@command(
-    discord=True,
-    allowed_roles = [ data["roles"]["admin"], data["roles"]["cmder"] ]
-)
-async def dcmd(args, msg, client):
-    cmd_name = utils.get(args, 0, "")
-    cmd = client.ccmds.get(cmd_name)
-    admin_role = client.main_guild.get_role(data["roles"]["admin"])
-    if not (cmd and (cmd["author"] == str(msg.author.id) or (admin_role in msg.author.roles))):
-        return await msg.reply(":x: You are not the author of the specified command or the command doesn't exist!")
-    client.ccmds.pop(cmd_name)
-    await client.update_ccmds()
-    await msg.reply(":white_check_mark: | Deleted the command")
-
-@command(
-    discord=True,
-    allowed_roles = [ data["roles"]["admin"], data["roles"]["cmder"] ]
-)
-async def ecmd(args, msg, client):
-    if len(args) < 3:
-        return await msg.reply(":x: Failed to edit the command. Please supply all the arguments\nFormat: `> ecmd <name> <compiler> <source>`")
-    name, compiler, source = args[0], args[1], args[2]
-    cmd = client.ccmds.get(name)
-
-    admin_role = client.main_guild.get_role(data["roles"]["admin"])
-    if not (cmd and (cmd["author"] == str(msg.author.id) or (admin_role in msg.author.roles))):
-        return await msg.reply(":x: You are not the author of the specified command or the command doesn't exist!")
-    
-
-    client.ccmds[name] = { "runner": compiler, "source": source, "author": str(msg.author.id) } # stringified id for backward compatibility
-    await client.update_ccmds()
-    await msg.reply(":white_check_mark: | Editted the command")
